requirements.txt
python-tds>=1.6
mssql-python>=0.1.0
pandas>=2.2.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
pyodbc>=5.1.0
streamlit>=1.31.0
python-dotenv>=1.0.0
pytest